requirements.txt
bs4==0.0.2
fastapi==0.116.1
langchain==0.3.27
langchain_chroma==0.2.5
langchain_community==0.3.27
langchain_core==0.3.74
langchain_huggingface==0.3.1
langchain_tavily==0.2.11
langchain_text_splitters==0.3.9
langchain_groq==0.3.6
langgraph==0.6.6
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
sentence_transformers==5.0.0
uvicorn==0.35.0